changed main.py and send.py
