+++ DataFetcher/main.py
@@ -1,30 +1,32 @@
 import datetime as dt
 import pandas as pd
 import pandas_datareader.data as web
+from recieve import recieve
 import urllib.request
+from yahoo_finance import Share
+import yfinance as yf
+import csv
+import zipfile
+import time
 
 def main():
-    url='https://s3.amazonaws.com/rawstore.datahub.io/b8d80d0c578007c9e0d199a6cd2625f1.csv'
+	
+    tickers = []
+    tickers = recieve()
+    timestr = time.strftime("%Y%m%d")
 
-    with urllib.request.urlopen(url) as testfile, open('dataset.csv', 'w') as f:
-        f.write(testfile.read().decode())
-
-
-    start = dt.datetime(2018, 1, 1)
-    end = dt.datetime(2018, 1, 2)
-
-    marketData = []
-
-    df = pd.read_csv('dataset.csv')
-    nasdaqSymbols = list(df['Symbol'])
-
-    for symbolCounter, symbol in enumerate(nasdaqSymbols):
+    for x in tickers:
         try:
-            stock = web.DataReader(nasdaqSymbols[symbolCounter], 'yahoo', start, end)
-            marketData.append(stock)
-            print(marketData)
+	    data = yf.download(x, period = "1d")
+	    z = zipfile.ZipFile('/Algo-Trader-Lean/Data/equity/usa/daily/'+x+'.zip', "a") 
+	    z.write(data.to_csv(timestr+'.csv'))
+	    z.close()
         except:
             continue
+
+
+
+
 
 if __name__ == "__main__":
     main()
